feed service & cleanup
