+++ packages/pds/src/api/app/bsky/feed/getAuthorFeed.ts
@@ -1,9 +1,8 @@
-import { sql } from 'kysely'
 import { Server } from '../../../../lexicon'
-import { FeedItemType, FeedKeyset, composeFeed } from '../util/feed'
-import { countAll } from '../../../../db/util'
+import { FeedKeyset, composeFeed } from '../util/feed'
 import { paginate } from '../../../../db/pagination'
 import AppContext from '../../../../context'
+import { FeedRow } from '../../../../services/feed'
 
 export default function (server: Server, ctx: AppContext) {
   server.app.bsky.feed.getAuthorFeed({
@@ -14,6 +13,8 @@ export default function (server: Server, ctx: AppContext) {
       const db = ctx.db.db
       const { ref } = db.dynamic
 
+      const feedService = ctx.services.feed(ctx.db, ctx.imgUriBuilder)
+
       const userLookupCol = author.startsWith('did:')
         ? 'did_handle.did'
         : 'did_handle.handle'
@@ -22,145 +23,42 @@ export default function (server: Server, ctx: AppContext) {
         .selectAll()
         .where(userLookupCol, '=', author)
 
-      const postsQb = db
-        .selectFrom('post')
+      const postsQb = feedService
+        .selectPostQb()
         .whereExists(
           userQb.whereRef('did_handle.did', '=', ref('post.creator')),
         )
-        .select([
-          sql<FeedItemType>`${'post'}`.as('type'),
-          'uri as postUri',
-          'cid as postCid',
-          'creator as originatorDid',
-          'indexedAt as cursor',
-        ])
 
-      const repostsQb = db
-        .selectFrom('repost')
+      const repostsQb = feedService
+        .selectRepostQb()
         .whereExists(
           userQb.whereRef('did_handle.did', '=', ref('repost.creator')),
         )
-        .select([
-          sql<FeedItemType>`${'repost'}`.as('type'),
-          'subject as postUri',
-          'subjectCid as postCid',
-          'creator as originatorDid',
-          'indexedAt as cursor',
-        ])
 
-      const trendsQb = db
-        .selectFrom('trend')
+      const trendsQb = feedService
+        .selectTrendQb()
         .whereExists(
           userQb.whereRef('did_handle.did', '=', ref('trend.creator')),
         )
-        .select([
-          sql<FeedItemType>`${'trend'}`.as('type'),
-          'subject as postUri',
-          'subjectCid as postCid',
-          'creator as originatorDid',
-          'indexedAt as cursor',
-        ])
-
-      let feedItemsQb = db
-        .selectFrom(postsQb.union(repostsQb).union(trendsQb).as('feed_items'))
-        .innerJoin('post', 'post.uri', 'postUri')
-        .innerJoin('ipld_block', 'ipld_block.cid', 'post.cid')
-        .innerJoin('did_handle as author', 'author.did', 'post.creator')
-        .leftJoin(
-          'profile as author_profile',
-          'author_profile.creator',
-          'author.did',
-        )
-        .innerJoin(
-          'did_handle as originator',
-          'originator.did',
-          'originatorDid',
-        )
-        .leftJoin(
-          'profile as originator_profile',
-          'originator_profile.creator',
-          'originatorDid',
-        )
-        .select([
-          'type',
-          'postUri',
-          'postCid',
-          'cursor',
-          'ipld_block.content as recordBytes',
-          'ipld_block.indexedAt as indexedAt',
-          'author.did as authorDid',
-          'author.declarationCid as authorDeclarationCid',
-          'author.actorType as authorActorType',
-          'author.handle as authorHandle',
-          'author.actorType as authorActorType',
-          'author_profile.displayName as authorDisplayName',
-          'author_profile.avatarCid as authorAvatarCid',
-          'originator.did as originatorDid',
-          'originator.declarationCid as originatorDeclarationCid',
-          'originator.actorType as originatorActorType',
-          'originator.handle as originatorHandle',
-          'originator.actorType as originatorActorType',
-          'originator_profile.displayName as originatorDisplayName',
-          'originator_profile.avatarCid as originatorAvatarCid',
-          db
-            .selectFrom('vote')
-            .whereRef('subject', '=', ref('postUri'))
-            .where('direction', '=', 'up')
-            .select(countAll.as('count'))
-            .as('upvoteCount'),
-          db
-            .selectFrom('vote')
-            .whereRef('subject', '=', ref('postUri'))
-            .where('direction', '=', 'down')
-            .select(countAll.as('count'))
-            .as('downvoteCount'),
-          db
-            .selectFrom('repost')
-            .whereRef('subject', '=', ref('postUri'))
-            .select(countAll.as('count'))
-            .as('repostCount'),
-          db
-            .selectFrom('post')
-            .whereRef('replyParent', '=', ref('postUri'))
-            .select(countAll.as('count'))
-            .as('replyCount'),
-          db
-            .selectFrom('repost')
-            .where('creator', '=', requester)
-            .whereRef('subject', '=', ref('postUri'))
-            .select('uri')
-            .as('requesterRepost'),
-          db
-            .selectFrom('vote')
-            .where('creator', '=', requester)
-            .whereRef('subject', '=', ref('postUri'))
-            .where('direction', '=', 'up')
-            .select('uri')
-            .as('requesterUpvote'),
-          db
-            .selectFrom('vote')
-            .where('creator', '=', requester)
-            .whereRef('subject', '=', ref('postUri'))
-            .where('direction', '=', 'down')
-            .select('uri')
-            .as('requesterDownvote'),
-        ])
 
       const keyset = new FeedKeyset(ref('cursor'), ref('postCid'))
+      let feedItemsQb = db
+        .selectFrom(postsQb.union(repostsQb).union(trendsQb).as('feed_items'))
+        .selectAll()
       feedItemsQb = paginate(feedItemsQb, {
         limit,
         before,
         keyset,
       })
 
-      const queryRes = await feedItemsQb.execute()
-      const feed = await composeFeed(db, ctx.imgUriBuilder, queryRes)
+      const feedItems: FeedRow[] = await feedItemsQb.execute()
+      const feed = await composeFeed(feedService, feedItems, requester)
 
       return {
         encoding: 'application/json',
         body: {
           feed,
-          cursor: keyset.packFromResult(queryRes),
+          cursor: keyset.packFromResult(feedItems),
         },
       }
     },
