+++ packages/pds/src/services/index.ts
@@ -1,8 +1,10 @@
 import { BlobStore } from '@atproto/repo'
 import Database from '../db'
 import { MessageQueue } from '../event-stream/types'
+import { ImageUriBuilder } from '../image/uri'
 import { ActorService } from './actor'
 import { AuthService } from './auth'
+import { FeedService } from './feed'
 import { RecordService } from './record'
 import { RepoService } from './repo'
 
@@ -13,6 +15,7 @@ export function createServices(
   return {
     actor: ActorService.creator(),
     auth: AuthService.creator(),
+    feed: FeedService.creator(),
     record: RecordService.creator(messageQueue),
     repo: RepoService.creator(messageQueue, blobstore),
   }
@@ -21,8 +24,13 @@ export function createServices(
 export type Services = {
   actor: FromDb<ActorService>
   auth: FromDb<AuthService>
+  feed: FromDbAndImgBuilder<FeedService>
   record: FromDb<RecordService>
   repo: FromDb<RepoService>
 }
 
 type FromDb<T> = (db: Database) => T
+type FromDbAndImgBuilder<T> = (
+  db: Database,
+  imgUriBuilder: ImageUriBuilder,
+) => T
